requirements.txt
streamlit
praw
google-generativeai
fpdf2